casts are necessary to build with gcc 3
git-svn-id: svn://svn.code.sf.net/p/marathon/code/trunk@2897 a75d40c9-0969-4759-85e5-97cb871b7ed4

+++ Source_Files/Misc/preferences_macintosh.h
@@ -186,7 +186,7 @@ const float SliderSensRange = 4;
 static float PrefToSlider(_fixed PrefVal)
 {
 	float CenteredSliderVal = (PrefVal > 0) ?
-		(log(float(PrefVal)) - log(FIXED_ONE))/log(SliderSensRange) :
+		(log(float(PrefVal)) - log((float) FIXED_ONE))/log(SliderSensRange) :
 		0;
 	
 	// -1 to 1 -> 0 to 1
